+++ lib/minitest/assert_dirs_equal.rb
@@ -21,4 +21,14 @@ class Minitest::Test
     matcher = AssertDirsEqual::Matcher.new(expected)
     refute matcher.matches?(target), matcher.failure_message_when_negated
   end
+
+  def assert_dir_included(expected, target)
+    matcher = AssertDirsEqual::Matcher.new(expected, exact_match: false)
+    assert matcher.matches?(target), matcher.failure_message
+  end
+
+  def refute_dir_included(expected, target)
+    matcher = AssertDirsEqual::Matcher.new(expected, exact_match: false)
+    refute matcher.matches?(target), matcher.failure_message_when_negated
+  end
 end