+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 ## Unreleased
 
+### Enhancements
+
+* Allow to compare dirs without exact match (i.e. there can be other files and directories inside
+`target_dir` that the matcher won't care about)
+
 ## 0.1.1
 
 ### Bug fixes